08 18 목요일 작업시작 전 commit
