+++ src/Routes.tsx
@@ -5,9 +5,10 @@ import AllCafesLocation from 'components/order/allcafe/AllCafesLocation';
 import AllDetails from 'components/order/allproduct/AllDetails';
 import AllProducts from 'components/order/allproduct/AllProducts';
 import Featured from 'components/order/featured/Featured';
-import OrderNow from 'components/order/featured/featuredList/OrederNow';
+import OrderNow from 'components/order/featured/OrederNow';
 import Order from 'components/order/Order';
 import FindStore from 'components/stores/FindStore';
+import Payment from 'components/order/featured/Payment';
 
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
@@ -24,6 +25,7 @@ const MainRoutes = () => {
           <Route path='cafes/menu/hot-coffees' element={ <AllDetails />}/>
           <Route path='featured' element={ <Featured /> } />
           <Route path='featured/order-now' element={<OrderNow/>}/>
+          <Route path='featured/order-now/payment' element={<Payment/>}/>
         </Route>
         <Route path='/store' element={ <FindStore /> } />
       </Routes>
