+++ src/components/order/featured/Featured.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route } from 'react-router-dom'
 import FeaturedMenu from './featuredList/FeaturedMenu'
-import OrderNow from './featuredList/OrederNow'
+import OrderNow from './OrederNow'
 
 const Featured = () => {
   return (
