+++ src/components/order/featured/Payment.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const payment = () => {
+  return (
+    <div>
+      
+    </div>
+  );
+};
+
+export default payment;
